+++ src/main/java/ru/nino/mybar/repository/DefaultRepository.java
@@ -3,6 +3,6 @@ package ru.nino.mybar.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.nino.mybar.entity.IdEntity;
 
-public interface DefaultRepository<ENTITY extends IdEntity> extends JpaRepository<ENTITY, Integer> {
+public interface DefaultRepository<ENTITY extends IdEntity> extends JpaRepository<ENTITY, Long> {
 
 }
